LayoutTests/fast/selectors/nth-child-of-with-invalid-An+B-expected.txt

Check the basic features of the ":nth-child(An+B of selectorList)" pseudo class.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


Testing ":nth-child(0 of testcaseA, testcaseB)"
PASS document.querySelectorAll('#test-root :nth-child(0 of testcaseA, testcaseB)').length is 0
PASS styledElements is []

Testing ":nth-child(0n of testcaseA, testcaseB)"
PASS document.querySelectorAll('#test-root :nth-child(0n of testcaseA, testcaseB)').length is 0
PASS styledElements is []

Testing ":nth-child(0n-1 of testcaseA, testcaseB)"
PASS document.querySelectorAll('#test-root :nth-child(0n-1 of testcaseA, testcaseB)').length is 0
PASS styledElements is []

Testing ":nth-child(-n+0 of testcaseA, testcaseB)"
PASS document.querySelectorAll('#test-root :nth-child(-n+0 of testcaseA, testcaseB)').length is 0
PASS styledElements is []

Testing ":nth-child(-1 of testcaseA, testcaseB)"
PASS document.querySelectorAll('#test-root :nth-child(-1 of testcaseA, testcaseB)').length is 0
PASS styledElements is []

Testing ":nth-child(-n-1 of testcaseA, testcaseB)"
PASS document.querySelectorAll('#test-root :nth-child(-n-1 of testcaseA, testcaseB)').length is 0
PASS styledElements is []

PASS successfullyParsed is true

TEST COMPLETE
